make sure that final install directory exists before trying to remove it

+++ easybuild/easyblocks/generic/binary.py
@@ -112,7 +112,8 @@ class Binary(EasyBlock):
             self.installdir = self.actual_installdir
             try:
                 # copytree expects target directory to not exist yet
-                shutil.rmtree(self.installdir)
+                if os.path.exists(self.installdir):
+                    shutil.rmtree(self.installdir)
                 shutil.copytree(staged_installdir, self.installdir)
             except OSError, err:
                 tup = (staged_installdir, self.installdir, err)
